move some of the Validators to an ExtraValidators file, and add compile flag to disable compilation of some of them.
