+++ tests/TransformTest.cpp
@@ -576,160 +576,6 @@ TEST_CASE_METHOD(TApp, "BoundTests", "[transform]") {
     CHECK(help.find("[3.4 - 5.9]") != std::string::npos);
 }
 
-TEST_CASE_METHOD(TApp, "NumberWithUnitCorrectlySplitNumber", "[transform]") {
-    std::map<std::string, int> mapping{{"a", 10}, {"b", 100}, {"cc", 1000}};
-
-    int value = 0;
-    app.add_option("-n", value)->transform(CLI::AsNumberWithUnit(mapping));
-
-    args = {"-n", "42"};
-    run();
-    CHECK(42 == value);
-
-    args = {"-n", "42a"};
-    run();
-    CHECK(420 == value);
-
-    args = {"-n", "  42  cc  "};
-    run();
-    CHECK(42000 == value);
-    args = {"-n", "  -42  cc  "};
-    run();
-    CHECK(-42000 == value);
-}
-
-TEST_CASE_METHOD(TApp, "NumberWithUnitFloatTest", "[transform]") {
-    std::map<std::string, double> mapping{{"a", 10}, {"b", 100}, {"cc", 1000}};
-    double value{0.0};
-    app.add_option("-n", value)->transform(CLI::AsNumberWithUnit(mapping));
-
-    args = {"-n", "42"};
-    run();
-    CHECK(42 == Approx(value));
-
-    args = {"-n", ".5"};
-    run();
-    CHECK(.5 == Approx(value));
-
-    args = {"-n", "42.5 a"};
-    run();
-    CHECK(425 == Approx(value));
-
-    args = {"-n", "42.cc"};
-    run();
-    CHECK(42000 == Approx(value));
-}
-
-TEST_CASE_METHOD(TApp, "NumberWithUnitCaseSensitive", "[transform]") {
-    std::map<std::string, int> mapping{{"a", 10}, {"A", 100}};
-
-    int value{0};
-    app.add_option("-n", value)->transform(CLI::AsNumberWithUnit(mapping, CLI::AsNumberWithUnit::CASE_SENSITIVE));
-
-    args = {"-n", "42a"};
-    run();
-    CHECK(420 == value);
-
-    args = {"-n", "42A"};
-    run();
-    CHECK(4200 == value);
-}
-
-TEST_CASE_METHOD(TApp, "NumberWithUnitCaseInsensitive", "[transform]") {
-    std::map<std::string, int> mapping{{"a", 10}, {"B", 100}};
-
-    int value{0};
-    app.add_option("-n", value)->transform(CLI::AsNumberWithUnit(mapping, CLI::AsNumberWithUnit::CASE_INSENSITIVE));
-
-    args = {"-n", "42a"};
-    run();
-    CHECK(420 == value);
-
-    args = {"-n", "42A"};
-    run();
-    CHECK(420 == value);
-
-    args = {"-n", "42b"};
-    run();
-    CHECK(4200 == value);
-
-    args = {"-n", "42B"};
-    run();
-    CHECK(4200 == value);
-}
-
-TEST_CASE_METHOD(TApp, "NumberWithUnitMandatoryUnit", "[transform]") {
-    std::map<std::string, int> mapping{{"a", 10}, {"A", 100}};
-
-    int value{0};
-    app.add_option("-n", value)
-        ->transform(CLI::AsNumberWithUnit(
-            mapping, CLI::AsNumberWithUnit::UNIT_REQUIRED | CLI::AsNumberWithUnit::CASE_SENSITIVE));
-
-    args = {"-n", "42a"};
-    run();
-    CHECK(420 == value);
-
-    args = {"-n", "42A"};
-    run();
-    CHECK(4200 == value);
-
-    args = {"-n", "42"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-}
-
-TEST_CASE_METHOD(TApp, "NumberWithUnitMandatoryUnit2", "[transform]") {
-    std::map<std::string, int> mapping{{"a", 10}, {"B", 100}};
-
-    int value{0};
-    app.add_option("-n", value)
-        ->transform(CLI::AsNumberWithUnit(
-            mapping, CLI::AsNumberWithUnit::UNIT_REQUIRED | CLI::AsNumberWithUnit::CASE_INSENSITIVE));
-
-    args = {"-n", "42A"};
-    run();
-    CHECK(420 == value);
-
-    args = {"-n", "42b"};
-    run();
-    CHECK(4200 == value);
-
-    args = {"-n", "42"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-}
-
-TEST_CASE_METHOD(TApp, "NumberWithUnitBadMapping", "[transform]") {
-    CHECK_THROWS_AS(CLI::AsNumberWithUnit(std::map<std::string, int>{{"a", 10}, {"A", 100}},
-                                          CLI::AsNumberWithUnit::CASE_INSENSITIVE),
-                    CLI::ValidationError);
-    CHECK_THROWS_AS(CLI::AsNumberWithUnit(std::map<std::string, int>{{"a", 10}, {"9", 100}}), CLI::ValidationError);
-    CHECK_THROWS_AS(CLI::AsNumberWithUnit(std::map<std::string, int>{{"a", 10}, {"AA A", 100}}), CLI::ValidationError);
-    CHECK_THROWS_AS(CLI::AsNumberWithUnit(std::map<std::string, int>{{"a", 10}, {"", 100}}), CLI::ValidationError);
-}
-
-TEST_CASE_METHOD(TApp, "NumberWithUnitBadInput", "[transform]") {
-    std::map<std::string, int> mapping{{"a", 10}, {"b", 100}};
-
-    int value{0};
-    app.add_option("-n", value)->transform(CLI::AsNumberWithUnit(mapping));
-
-    args = {"-n", "13 a b"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-    args = {"-n", "13 c"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-    args = {"-n", "a"};
-    // Assume 1.0 unit
-    CHECK_NOTHROW(run());
-    args = {"-n", "12.0a"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-    args = {"-n", "a5"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-    args = {"-n", ""};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-    args = {"-n", "13 a-"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-}
-
 static const std::map<std::string, std::string> validValues = {
     {"test\\u03C0\\u00e9", from_u8string(u8"test\u03C0\u00E9")},
     {"test\\u03C0\\u00e9", from_u8string(u8"test\u73C0\u0057")},
@@ -777,266 +623,6 @@ TEST_CASE_METHOD(TApp, "StringEscapeInvalid", "[transform]") {
     CHECK_THROWS_AS(run(), CLI::ValidationError);
 }
 
-TEST_CASE_METHOD(TApp, "NumberWithUnitIntOverflow", "[transform]") {
-    std::map<std::string, int> mapping{{"a", 1000000}, {"b", 100}, {"c", 101}};
 
-    std::int32_t value = 0;
-    app.add_option("-n", value)->transform(CLI::AsNumberWithUnit(mapping));
 
-    args = {"-n", "1000 a"};
-    run();
-    CHECK(1000000000 == value);
 
-    args = {"-n", "1000000 a"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-
-    args = {"-n", "-1000000 a"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-
-    args = {"-n", "21474836 b"};
-    run();
-    CHECK(2147483600 == value);
-
-    args = {"-n", "21474836 c"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-}
-
-TEST_CASE_METHOD(TApp, "NumberWithUnitFloatOverflow", "[transform]") {
-    std::map<std::string, float> mapping{{"a", 2.f}, {"b", 1.f}, {"c", 0.f}};
-
-    float value{0.0F};
-    app.add_option("-n", value)->transform(CLI::AsNumberWithUnit(mapping));
-
-    args = {"-n", "3e+38 a"};
-    CHECK_THROWS_AS(run(), CLI::ValidationError);
-
-    args = {"-n", "3e+38 b"};
-    run();
-    CHECK(3e+38f == Approx(value));
-
-    args = {"-n", "3e+38 c"};
-    run();
-    CHECK(0.f == Approx(value));
-}
-
-TEST_CASE_METHOD(TApp, "AsSizeValue1000_1024", "[transform]") {
-    std::uint64_t value{0};
-    app.add_option("-s", value)->transform(CLI::AsSizeValue(true));
-
-    args = {"-s", "10240"};
-    run();
-    CHECK(10240u == value);
-
-    args = {"-s", "1b"};
-    run();
-    CHECK(1u == value);
-
-    std::uint64_t k_value{1000u};
-    std::uint64_t ki_value{1024u};
-    args = {"-s", "1k"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1kb"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1 Kb"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1ki"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1kib"};
-    run();
-    CHECK(ki_value == value);
-
-    k_value = 1000ull * 1000u;
-    ki_value = 1024ull * 1024u;
-    args = {"-s", "1m"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1mb"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1mi"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1mib"};
-    run();
-    CHECK(ki_value == value);
-
-    k_value = 1000ull * 1000u * 1000u;
-    ki_value = 1024ull * 1024u * 1024u;
-    args = {"-s", "1g"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1gb"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1gi"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1gib"};
-    run();
-    CHECK(ki_value == value);
-
-    k_value = 1000ull * 1000u * 1000u * 1000u;
-    ki_value = 1024ull * 1024u * 1024u * 1024u;
-    args = {"-s", "1t"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1tb"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1ti"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1tib"};
-    run();
-    CHECK(ki_value == value);
-
-    k_value = 1000ull * 1000u * 1000u * 1000u * 1000u;
-    ki_value = 1024ull * 1024u * 1024u * 1024u * 1024u;
-    args = {"-s", "1p"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1pb"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1pi"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1pib"};
-    run();
-    CHECK(ki_value == value);
-
-    k_value = 1000ull * 1000u * 1000u * 1000u * 1000u * 1000u;
-    ki_value = 1024ull * 1024u * 1024u * 1024u * 1024u * 1024u;
-    args = {"-s", "1e"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1eb"};
-    run();
-    CHECK(k_value == value);
-    args = {"-s", "1ei"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1eib"};
-    run();
-    CHECK(ki_value == value);
-}
-
-TEST_CASE_METHOD(TApp, "duration_test", "[transform]") {
-    std::chrono::seconds duration{1};
-
-    app.option_defaults()->ignore_case();
-    app.add_option_function<std::size_t>(
-           "--duration",
-           [&](size_t a_value) { duration = std::chrono::seconds{a_value}; },
-           "valid units: sec, min, h, day.")
-        ->capture_default_str()
-        ->transform(CLI::AsNumberWithUnit(
-            std::map<std::string, std::size_t>{{"sec", 1}, {"min", 60}, {"h", 3600}, {"day", 24 * 3600}}));
-    CHECK_NOTHROW(app.parse(std::vector<std::string>{"1 day", "--duration"}));
-
-    CHECK(std::chrono::seconds(86400) == duration);
-}
-
-TEST_CASE_METHOD(TApp, "AsSizeValue1024", "[transform]") {
-    std::uint64_t value{0};
-    app.add_option("-s", value)->transform(CLI::AsSizeValue(false));
-
-    args = {"-s", "10240"};
-    run();
-    CHECK(10240u == value);
-
-    args = {"-s", "1b"};
-    run();
-    CHECK(1u == value);
-
-    std::uint64_t ki_value{1024u};
-    args = {"-s", "1k"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1kb"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1 Kb"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1ki"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1kib"};
-    run();
-    CHECK(ki_value == value);
-
-    ki_value = 1024ull * 1024u;
-    args = {"-s", "1m"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1mb"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1mi"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1mib"};
-    run();
-    CHECK(ki_value == value);
-
-    ki_value = 1024ull * 1024u * 1024u;
-    args = {"-s", "1g"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1gb"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1gi"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1gib"};
-    run();
-    CHECK(ki_value == value);
-
-    ki_value = 1024ull * 1024u * 1024u * 1024u;
-    args = {"-s", "1t"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1tb"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1ti"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1tib"};
-    run();
-    CHECK(ki_value == value);
-
-    ki_value = 1024ull * 1024u * 1024u * 1024u * 1024u;
-    args = {"-s", "1p"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1pb"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1pi"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1pib"};
-    run();
-    CHECK(ki_value == value);
-
-    ki_value = 1024ull * 1024u * 1024u * 1024u * 1024u * 1024u;
-    args = {"-s", "1e"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1eb"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1ei"};
-    run();
-    CHECK(ki_value == value);
-    args = {"-s", "1eib"};
-    run();
-    CHECK(ki_value == value);
-}
